# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Business Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /items:
    get:
      operationId: listItems
      tags: [items]
      summary: List all items
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
    post:
      operationId: createItem
      tags: [items]
      summary: Create item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
  /items/{id}:
    get:
      operationId: getItem
      tags: [items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
    patch:
      operationId: updateItem
      tags: [items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
    delete:
      operationId: deleteItem
      tags: [items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── INVENTORY ──────────────────────────────────────────────────────────────
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    patch:
      operationId: updateInvoice
      tags: [invoices]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /invoices/{id}/status:
    patch:
      operationId: updateInvoiceStatus
      tags: [invoices]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceStatusUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"

  # ── PURCHASE ORDERS ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Item:
      type: object
      required: [id, name, category, unitPrice, costPrice, unitOfMeasure, createdAt]
      properties:
        id:            { type: integer }
        name:          { type: string }
        description:   { type: ["string", "null"] }
        category:      { type: string }
        unitPrice:     { type: number }
        costPrice:     { type: number }
        unitOfMeasure: { type: string }
        createdAt:     { type: string }
        updatedAt:     { type: string }
    ItemInput:
      type: object
      required: [name, category, unitPrice, costPrice, unitOfMeasure]
      properties:
        name:          { type: string }
        description:   { type: string }
        category:      { type: string }
        unitPrice:     { type: number }
        costPrice:     { type: number }
        unitOfMeasure: { type: string }
    ItemUpdate:
      type: object
      properties:
        name:          { type: string }
        description:   { type: string }
        category:      { type: string }
        unitPrice:     { type: number }
        costPrice:     { type: number }
        unitOfMeasure: { type: string }

    # ── INVENTORY ─────────────────────────────────────────────────────────────
...
    Customer:
      type: object
      required: [id, name, createdAt]
      properties:
        id:        { type: integer }
        name:      { type: string }
        email:     { type: ["string", "null"] }
        phone:     { type: ["string", "null"] }
        address:   { type: ["string", "null"] }
        city:      { type: ["string", "null"] }
        country:   { type: ["string", "null"] }
        notes:     { type: ["string", "null"] }
        createdAt: { type: string }
...
    InvoiceLineItem:
      type: object
      required: [id, invoiceId, itemId, description, quantity, unitPrice, lineTotal]
      properties:
        id:          { type: integer }
        invoiceId:   { type: integer }
        itemId:      { type: ["integer", "null"] }
        description: { type: string }
        quantity:    { type: number }
        unitPrice:   { type: number }
        lineTotal:   { type: number }
        item:
          $ref: "#/components/schemas/Item"
    Invoice:
      type: object
      required: [id, invoiceNumber, customerId, status, subtotal, taxRate, taxAmount, total, createdAt]
      properties:
        id:            { type: integer }
        invoiceNumber: { type: string }
        customerId:    { type: integer }
        status:        { type: string, enum: [Draft, Sent, Paid, Overdue] }
        issueDate:     { type: ["string", "null"] }
        dueDate:       { type: ["string", "null"] }
        subtotal:      { type: number }
        taxRate:       { type: number }
        taxAmount:     { type: number }
        total:         { type: number }
        notes:         { type: ["string", "null"] }
        createdAt:     { type: string }
        customer:
          $ref: "#/components/schemas/Customer"
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceLineItem"
    InvoiceLineItemInput:
      type: object
      required: [description, quantity, unitPrice]
      properties:
        itemId:      { type: integer }
        description: { type: string }
        quantity:    { type: number }
        unitPrice:   { type: number }
    InvoiceInput:
      type: object
      required: [customerId, lineItems]
      properties:
        customerId: { type: integer }
        issueDate:  { type: string }
        dueDate:    { type: string }
        taxRate:    { type: number }
        notes:      { type: string }
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceLineItemInput"
    InvoiceUpdate:
      type: object
      properties:
        customerId: { type: integer }
        issueDate:  { type: string }
        dueDate:    { type: string }
        taxRate:    { type: number }
        notes:      { type: string }
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceLineItemInput"
    InvoiceStatusUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [Draft, Sent, Paid, Overdue] }

    # ── PURCHASE ORDER ────────────────────────────────────────────────────────
...