# lib/api-zod/src/generated/types/invoiceUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Business Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { InvoiceLineItemInput } from './invoiceLineItemInput';

export interface InvoiceUpdate {
  customerId?: number;
  issueDate?: string;
  dueDate?: string;
  taxRate?: number;
  notes?: string;
  lineItems?: InvoiceLineItemInput[];
}