# lib/api-zod/src/generated/types/itemUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Business Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface ItemUpdate {
  name?: string;
  description?: string;
  category?: string;
  unitPrice?: number;
  costPrice?: number;
  unitOfMeasure?: string;
}